Feat: Divide serializer into multiple files in 'user' app
