+++ diningcoach/user/serializers/account_serializers.py
@@ -149,35 +149,3 @@ class UserLoginSerializer(serializers.Serializer):
 
     return data
   '''
-
-
-class UserDefaultSerializer(serializers.ModelSerializer):
-  class Meta:
-    model = User
-    exclude = ['password', 'is_staff', 'is_superuser', 'groups', 'user_permissions']
-
-
-class UserProfileDefaultSerializer(serializers.ModelSerializer):
-  class Meta:
-    model = UserProfile
-    fields = '__all__'
-
-
-class UserHealthDefaultSerializer(serializers.ModelSerializer):
-  class Meta:
-    model = UserHealth
-    fields = '__all__'
-
-
-class UserProfileRetrieveSerializer(serializers.ModelSerializer):
-  profile_info = UserProfileDefaultSerializer(many=False, read_only=True)
-
-  class Meta(UserDefaultSerializer.Meta):
-    pass
-
-
-class UserHealthRetrieveSerializer(serializers.ModelSerializer):
-  health_info = UserHealthDefaultSerializer(many=False, read_only=True)
-
-  class Meta(UserDefaultSerializer.Meta):
-    pass
