+++ diningcoach/user/serializers/platform_serializers.py
@@ -1,0 +1,19 @@
+from user.models import User, UserProfile, UserHealth
+
+from rest_framework import serializers
+
+
+class AuthUserKakaoSerializer(serializers.Serializer):
+  pass
+
+
+class AuthUserGoogleSerializer(serializers.Serializer):
+  pass
+
+
+class AuthUserAppleSerializer(serializers.Serializer):
+  pass
+
+
+class AuthUserNaverSerializer(serializers.Serializer):
+  pass
